requirements.txt
black==23.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
idna==3.4
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pycparser==2.21
PySoundFile==0.9.0.post1
pyTelegramBotAPI==4.9.0
python-dotenv==0.21.1
requests==2.31.0
SpeechRecognition==3.10.0
urllib3==2.0.3
